requirements.txt
agno==1.0.8
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2pdf==0.1.8
duckduckgo_search==7.3.2
exceptiongroup==1.2.2
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.2
openai==1.61.1
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
primp==0.12.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
yfinance==0.2.52
